+++ Course/Course/Presentation/Unit/CourseUnitView.swift
@@ -38,7 +38,7 @@ public struct CourseUnitView: View {
     @State private var videoURLs: [String: URL?] = [:]
     @State private var webURLs: [String: URL?] = [:]
 
-    @State private var uiScrollView: UIScrollView?
+    @State var currentBlock: CourseBlock?
 
     let isDropdownActive: Bool
     
@@ -176,26 +176,16 @@ public struct CourseUnitView: View {
 
         UnitStack(isVerticalNavigation: !isHorizontalNavigation, alignment: alignment, spacing: 0) {
             let data = Array(viewModel.verticals[viewModel.verticalIndex].childs.enumerated())
-            ForEach(data, id: \.offset) { index, block in
+            ForEach(data, id: \.offset) {index, block in
                 VStack(spacing: 0) {
-
+                    
                     if viewModel.showVideoNavigation {
                         if !isHorizontal {
-                            videoNavigationView(block: block)
-                                .padding(.bottom, 17)
-                            
-                            Divider()
-                                .padding(.bottom, 16)
-                            
-                            VStack(alignment: .leading) {
-                                Text(viewModel.createBreadCrumpsForVideoNavigation(video: block))
-                                    .font(Theme.Fonts.bodySmall)
-                                    .foregroundStyle(Theme.Colors.textPrimary)
-                                    .multilineTextAlignment(.leading)
-                                    .lineLimit(2)
-                                    .padding(.bottom, 8)
-                                    .padding(.horizontal, 16)
-                            }
+                            VideoNavigationView(
+                                viewModel: viewModel,
+                                currentBlock: $currentBlock,
+                                block: block
+                            )
                         }
                     }
 
@@ -235,8 +225,6 @@ public struct CourseUnitView: View {
             }
         }
     }
-
-    @State var currentBlock: CourseBlock?
 
     @ViewBuilder
     private func contentView(for block: CourseBlock, index: Int, reader: GeometryProxy) -> some View {
@@ -295,64 +283,6 @@ public struct CourseUnitView: View {
                 index: index,
                 reader: reader
             )
-        }
-    }
-
-    @ViewBuilder
-    private func videoNavigationView(block: CourseBlock) -> some View {
-        ScrollViewReader { scrollProxy in
-            ScrollView(.horizontal, showsIndicators: false) {
-                LazyHStack(spacing: 0) {
-                    ForEach(Array(viewModel.allVideosForNavigation.enumerated()), id: \.element.id) { index, video in
-                        VideoThumbnailView(
-                            thumbnailData: VideoThumbnailData(
-                                video: video,
-                                chapter: nil,
-                                courseStructure: nil,
-                                onVideoTap: { video, _ in
-                                    viewModel.handleVideoTap(video: video)
-                                }
-                            ),
-                            thumbnailWidth: 122,
-                            thumbnailHeight: 67,
-                            isCurrentVideo: video == currentBlock
-                        )
-                        .padding(.leading, index == 0 ? 24 : 8)
-                        .id(video.id)
-
-                        if index == viewModel.allVideosForNavigation.count - 1 {
-                            Spacer(minLength: 100)
-                        }
-                    }
-                }
-            }
-            .introspect(.scrollView, on: .iOS(.v16, .v17, .v18)) { scroll in
-                DispatchQueue.main.async {
-                    uiScrollView = scroll
-                }
-            }
-            .onChange(of: viewModel.allVideosForNavigation) { _ in
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                    scrollToCurrentVide(block: block) { id in
-                        scrollProxy.scrollTo(id, anchor: .leading)
-                    }
-                }
-            }
-            .frame(height: 72)
-        }
-    }
-
-    private func scrollToCurrentVide(block: CourseBlock, scrollTo: @escaping (String) -> Void) {
-        guard let currentVideo = viewModel.allVideosForNavigation.first(where: { $0 == block }) else {
-            return
-        }
-
-        scrollTo(currentVideo.id)
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-            guard let scroll = uiScrollView else { return }
-            let newX = max(scroll.contentOffset.x - 20, 0)
-            scroll.setContentOffset(CGPoint(x: newX, y: 0), animated: true)
         }
     }
